requirements.txt
Jinja2==2.9.5
mock==2.0.0
pytest==3.2.0
pytest-cov==2.5.1
#pyspark==2.2.0
google-cloud-bigquery==0.26.0
google-cloud-storage==1.3.1
numpy==1.12.1
